+++ chatkit-android/src/main/kotlin/com/pusher/chatkit/BeamsPushNotifications.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import com.pusher.chatkit.pushnotifications.BeamsTokenProviderService
 import com.pusher.chatkit.pushnotifications.PushNotifications
 import com.pusher.chatkit.util.FutureValue
-import com.pusher.pushnotifications.Callback
+import com.pusher.pushnotifications.BeamsCallback
 import com.pusher.pushnotifications.PusherCallbackError
 import com.pusher.pushnotifications.auth.TokenProvider
 import com.pusher.util.Result
@@ -18,15 +18,14 @@ class BeamsPushNotifications(
         private val beamsTokenProviderService: BeamsTokenProviderService
 ) : PushNotifications {
 
-    override fun start(): Result<Unit, Error> {
-        val tokenProvider = object : TokenProvider {
-            override fun fetchToken(userId: String): String? {
-                return beamsTokenProviderService.fetchToken(userId).map { it.token }.recover { null }
-            }
-        }
+    private val tokenProvider = object : TokenProvider {
+        override fun fetchToken(userId: String): String =
+                beamsTokenProviderService.fetchToken(userId)
+    }
 
+    override fun start(): Result<Unit, Error> {
         return try {
-            Beams.start(context, instanceId, tokenProvider)
+            Beams.start(context, instanceId)
             Result.success(Unit)
         } catch (ex: ClassNotFoundException) {
             Result.failure(elements.OtherError("It seems like some dependencies that Push Notifications requires are missing. Check https://docs.pusher.com/chatkit for more information.", ex.cause))
@@ -38,7 +37,7 @@ class BeamsPushNotifications(
     override fun setUserId(userId: String): Result<Unit, Error> {
         val f = FutureValue<Result<Unit, Error>>()
         try {
-            Beams.setUserId(userId, object : Callback<Void, PusherCallbackError> {
+            Beams.setUserId(userId, tokenProvider, object : BeamsCallback<Void, PusherCallbackError> {
                 override fun onFailure(error: PusherCallbackError) {
                     f.set(Result.failure(elements.OtherError(error.message, error.cause)))
                 }
@@ -57,23 +56,13 @@ class BeamsPushNotifications(
     }
 
     override fun stop(): Result<Unit, Error> {
-        val f = FutureValue<Result<Unit, Error>>()
         try {
-            Beams.stop(object : Callback<Void, PusherCallbackError> {
-                override fun onFailure(error: PusherCallbackError) {
-                    f.set(Result.failure(elements.OtherError(error.message, error.cause)))
-                }
-
-                override fun onSuccess(vararg values: Void) {
-                    f.set(Result.success(Unit))
-                }
-            })
+            Beams.stop()
         } catch (ex: ClassNotFoundException) {
             return Result.failure(elements.OtherError("It seems like some dependencies that Push Notifications requires are missing. Check https://docs.pusher.com/chatkit for more information.", ex.cause))
         } catch (ex: Throwable) {
             return Result.failure(elements.OtherError(ex.message ?: "Unknown error", ex.cause))
         }
-
-        return f.get()
+        return Result.success(Unit)
     }
 }